+++ resources/aws_apigateway_v2_deployment.rb
@@ -5,7 +5,7 @@ resource_name :aws_apigateway_v2_deployment
 provides :aws_apigateway_v2_deployment, target_mode: true, platform: "aws"
 
 description <<~DESCRIPTION
-  Resource Type definition for AWS::ApiGatewayV2::Deployment
+  The ``AWS::ApiGatewayV2::Deployment`` resource creates a deployment for an API.
 DESCRIPTION
 
 property :name, String,
@@ -17,19 +17,25 @@ property :api_id, String,
          callbacks: {
            "api_id is not a String" => lambda { |v| v.is_a? String },
          },
-         description: ""
+         description: <<~'DESCRIPTION'
+           The API identifier.
+         DESCRIPTION
 
 property :description, String,
          callbacks: {
            "description is not a String" => lambda { |v| v.is_a? String },
          },
-         description: ""
+         description: <<~'DESCRIPTION'
+           The description for the deployment resource.
+         DESCRIPTION
 
 property :stage_name, String,
          callbacks: {
            "stage_name is not a String" => lambda { |v| v.is_a? String },
          },
-         description: ""
+         description: <<~'DESCRIPTION'
+           The name of an existing stage to associate with the deployment.
+         DESCRIPTION
 
 # API URLs and mappings
 rest_api_collection "/AWS::ApiGatewayV2::Deployment"