+++ resources/aws_apigateway_v2_model.rb
@@ -5,7 +5,7 @@ resource_name :aws_apigateway_v2_model
 provides :aws_apigateway_v2_model, target_mode: true, platform: "aws"
 
 description <<~DESCRIPTION
-  Resource Type definition for AWS::ApiGatewayV2::Model
+  The ``AWS::ApiGatewayV2::Model`` resource updates data model for a WebSocket API. For more information, see [Model Selection Expressions](https://docs.aws.amazon.com/apigateway/latest/developerguide/apigateway-websocket-api-selection-expressions.html#apigateway-websocket-api-model-selection-expressions) in the *API Gateway Developer Guide*.
 DESCRIPTION
 
 property :name, String,
@@ -17,19 +17,25 @@ property :api_id, String,
          callbacks: {
            "api_id is not a String" => lambda { |v| v.is_a? String },
          },
-         description: ""
+         description: <<~'DESCRIPTION'
+           The API identifier.
+         DESCRIPTION
 
 property :content_type, String,
          callbacks: {
            "content_type is not a String" => lambda { |v| v.is_a? String },
          },
-         description: ""
+         description: <<~'DESCRIPTION'
+           The content-type for the model, for example, "application/json".
+         DESCRIPTION
 
 property :description, String,
          callbacks: {
            "description is not a String" => lambda { |v| v.is_a? String },
          },
-         description: ""
+         description: <<~'DESCRIPTION'
+           The description of the model.
+         DESCRIPTION
 
 property :name, String,
          name_property: true,
@@ -37,14 +43,18 @@ property :name, String,
          callbacks: {
            "name is not a String" => lambda { |v| v.is_a? String },
          },
-         description: ""
+         description: <<~'DESCRIPTION'
+           The name of the model.
+         DESCRIPTION
 
 property :schema, Hash,
          required: true,
          callbacks: {
            "schema is not a Object" => lambda { |v| v.is_a? Object },
          },
-         description: ""
+         description: <<~'DESCRIPTION'
+           The schema for the model. For application/json models, this should be JSON schema draft 4 model.
+         DESCRIPTION
 
 # API URLs and mappings
 rest_api_collection "/AWS::ApiGatewayV2::Model"